+++ app/main/redirects.py
@@ -30,6 +30,14 @@ def guests() -> Response:
     return redirect_url(url_for("guests.index"))
 
 
+@blueprint.route("/guest/<string:guest_slug>")
+def guests_details(guest_slug: str) -> Response:
+    """Redirect: /guest/<guest_slug> to /guests/<guest_slug>."""
+    return redirect_url(
+        url_for("guests.details", guest_slug=guest_slug), status_code=301
+    )
+
+
 @blueprint.route("/help")
 def help_page() -> Response:
     """Redirect: /help to /."""
@@ -42,6 +50,12 @@ def help_page() -> Response:
 def hosts() -> Response:
     """Redirect: /host and /hosts to /hosts/."""
     return redirect_url(url_for("hosts.index"))
+
+
+@blueprint.route("/host/<string:host_slug>")
+def hosts_details(host_slug: str) -> Response:
+    """Redirect: /host/<host_slug> to /hosts/<host_slug>."""
+    return redirect_url(url_for("hosts.details", host_slug=host_slug), status_code=301)
 
 
 @blueprint.route("/info")
@@ -58,6 +72,14 @@ def locations() -> Response:
     return redirect_url(url_for("locations.index"))
 
 
+@blueprint.route("/location/<string:location_slug>")
+def locations_details(location_slug: str) -> Response:
+    """Redirect: /location/<location_slug> to /locations/<location_slug>."""
+    return redirect_url(
+        url_for("locations.details", location_slug=location_slug), status_code=301
+    )
+
+
 @blueprint.route("/panelist")
 @blueprint.route("/panelist/")
 @blueprint.route("/panelists")
@@ -66,12 +88,29 @@ def panelists() -> Response:
     return redirect_url(url_for("panelists.index"))
 
 
+@blueprint.route("/panelist/<string:panelist_slug>")
+def panelists_details(panelist_slug: str) -> Response:
+    """Redirect: /panelist/<panelist_slug> to /panelists/<panelist_slug>."""
+    return redirect_url(
+        url_for("panelists.details", panelist_slug=panelist_slug), status_code=301
+    )
+
+
 @blueprint.route("/scorekeeper")
 @blueprint.route("/scorekeeper/")
 @blueprint.route("/scorekeepers")
 def scorekeepers() -> Response:
     """Redirect: /scorekeeper and /scorekeepers to /scorekeepers/."""
     return redirect_url(url_for("scorekeepers.index"))
+
+
+@blueprint.route("/scorekeeper/<string:scorekeeper_slug>")
+def scorekeepers_details(scorekeeper_slug: str) -> Response:
+    """Redirect: /scorekeeper/<scorekeeper_slug> to /scorekeepers/<scorekeeper_slug>."""
+    return redirect_url(
+        url_for("scorekeepers.details", scorekeeper_slug=scorekeeper_slug),
+        status_code=301,
+    )
 
 
 @blueprint.route("/search")
@@ -86,6 +125,43 @@ def search() -> Response:
 def shows() -> Response:
     """Redirect: /show and /shows to /shows/."""
     return redirect_url(url_for("shows.index"))
+
+
+@blueprint.route("/show/<string:iso_date_string>")
+def shows_date_string(iso_date_string: str) -> Response:
+    """Redirect: /show/<iso_date_string_year> to /shows/<iso_date_string>."""
+    return redirect_url(
+        url_for("shows.date_string", iso_date_string=iso_date_string), status_code=301
+    )
+
+
+@blueprint.route("/show/<int:show_year>")
+def shows_year(show_year: int) -> Response:
+    """Redirect: /show/<show_year> to /shows/<show_year>."""
+    return redirect_url(url_for("shows.year", show_year=show_year), status_code=301)
+
+
+@blueprint.route("/show/<int:show_year>/<int:show_month>")
+def shows_year_month(show_year: int, show_month: int) -> Response:
+    """Redirect: /show/<show_year>/<show_month> to /shows/<show_year>/<show_month>."""
+    return redirect_url(
+        url_for("shows.year_month", show_year=show_year, show_month=show_month),
+        status_code=301,
+    )
+
+
+@blueprint.route("/show/<int:show_year>/<int:show_month>/<int:show_day>")
+def shows_year_month_day(show_year: int, show_month: int, show_day: int) -> Response:
+    """Redirect: /show/<show_year>/<show_month>/<show_day> to /shows/<show_year>/<show_month>/<show_day>."""
+    return redirect_url(
+        url_for(
+            "shows.year_month_day",
+            show_year=show_year,
+            show_month=show_month,
+            show_day=show_day,
+        ),
+        status_code=301,
+    )
 
 
 @blueprint.route("/shows/best-of-repeats")